Corrected categories generation
Corrected a bug as some category numbers don't exist (for example, 45), this caused a bug where if a card had category 45 in his `categoriesSum`, all categories in CATEGORIES from 44 to 1 are added too. So the code shouldn't check for every existing category in CATEGORIES, but for every possible numbers
Removed the power of CATEGORIES indexes, useless now.